backend/commons/runtime/runtime-docker/src/test/java/com/kraken/runtime/docker/DockerTaskServiceTest.java
package com.kraken.runtime.docker;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
import org.springframework.boot.test.context.ConfigFileApplicationContextInitializer;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringRunner;

import static org.assertj.core.api.Assertions.assertThat;

@RunWith(SpringRunner.class)
@ContextConfiguration(classes = {TestConfiguration.class}, initializers = {ConfigFileApplicationContextInitializer.class})
@EnableAutoConfiguration
public class DockerTaskServiceTest {

  @Autowired
  DockerTaskService taskService;

  @Test
  public void shouldCountHosts() {
    assertThat(taskService.hostsCount().block()).isEqualTo(1);
  }

}
